Lookup object and error are now in dedicated files.
This is to share them with avax implementation.
