+++ models/lookup-object.ts
@@ -1,0 +1,7 @@
+interface LookupObject {
+  name: string;
+  phone: string;
+  address: string;
+}
+
+export default LookupObject;
